+++ src/animal/Wolf.cpp
@@ -12,7 +12,7 @@ Wolf::Wolf(uint16_t x, uint16_t y, World *world) : Animal(x, y, world) {
     initiative = 5;
 }
 
-Wolf::Wolf(uint16_t x, uint16_t y, uint8_t power, uint8_t initiative, World *world)
+Wolf::Wolf(uint16_t x, uint16_t y, uint8_t power, uint16_t initiative, World *world)
         : Animal(x, y, power, initiative, world) {
     sign = 'W';
     this->initiative = 5;
